fivestars/requirements.txt
django == 1.10
PyMySQL
djangorestframework
requests